Fixed race condition when notifying HTTP/2 `HeadersFrame`s.
Before, `Stream.Listener.onHeaders()` was assuming that the headers were processed synchronously.
Now, `HttpReceiverOverHTTP2` process them asynchronously, with a task that declares an invocation type.
This was causing a race between the task and the code present after the call to `onHeaders()`.

Introduced `Stream.Listener.onHeaders(Stream, HeadersFrame, Callback)` to allow asynchronous processing of `HeadersFrame`s.

Optimised `HttpReceiver.responseHeaders()` to avoid creating the `Content.Source` if there is no content.

Fixed `IteratingCallback` and `HTTP2Flusher` to use `tryLock()` in `toString()` to avoid deadlocks.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>