Partial fix for #12681 caching content
This is a partial fix for #12681. It does not fix the actual release race on a single content, but by restricting the shrinking to a single thread, make the race less likely to occur.

There is a more comprehensive fix in #12678 for 12.1, but the cache has changed too much for a straight forward back port.